+++ src/pages/aql/game_id.tsx
