+++ src/pages/game_id/board.tsx
