+++ src/pages/game_id/config.tsx
